Undo folder rename as it was causing a different doc manifest
